Refactor: patient-option.class -> toggle.class
